Day seventeen: brute force is good actually
Spent a lot of time pursuing an analytic solution to
this one. I'm still sure it's possible, but I ran into
a major snag after a few days of thinking & tinkering
and decided to just do it the straightforward way.

Oh and I changed Max to take varargs instead of a slice.